requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
Email-validator==2.1.0.post1
python-dotenv==1.0.0
flask-sock==0.6.0
flask-migrate==4.0.5
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-marshmallow==1.1.0
marshmallow-sqlalchemy==1.0.0
pytz==2024.1
WTForms==3.1.2
SQLAlchemy==2.0.25
click==8.1.7
Jinja2==3.1.3
itsdangerous==2.1.2
MarkupSafe==2.1.4
packaging==23.2
typing_extensions==4.9.0
paho-mqtt==2.0.0
waitress==2.1.2